TARGET/requirement.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0  # PostgreSQL
aiomysql>=0.2.0  # MySQL/MariaDB
pymysql>=1.1.0

# AI/LLM
openai>=1.0.0
anthropic>=0.8.0
aiohttp>=3.9.0  # For Ollama/webhooks

# Export
openpyxl>=3.1.0  # Excel
weasyprint>=60.0  # PDF
python-multipart>=0.0.6

# Email
aiosmtplib>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
